Move _errors file up to the top level.
Plan to raise more subtypes of DeviceError that have to do with Monitor
and so forth.
