+++ src/pyudev/__init__.py
@@ -39,16 +39,16 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
+from pyudev._errors import DeviceNotFoundAtPathError
+from pyudev._errors import DeviceNotFoundByFileError
+from pyudev._errors import DeviceNotFoundByNameError
+from pyudev._errors import DeviceNotFoundByNumberError
+from pyudev._errors import DeviceNotFoundError
+from pyudev._errors import DeviceNotFoundInEnvironmentError
 
 from pyudev.device import Attributes
 from pyudev.device import Device
 from pyudev.device import Devices
-from pyudev.device import DeviceNotFoundAtPathError
-from pyudev.device import DeviceNotFoundByFileError
-from pyudev.device import DeviceNotFoundByNameError
-from pyudev.device import DeviceNotFoundByNumberError
-from pyudev.device import DeviceNotFoundError
-from pyudev.device import DeviceNotFoundInEnvironmentError
 from pyudev.device import Tags
 
 from pyudev.discover import DeviceFileHypothesis
