+++ src/pyudev/_errors.py
