+++ src/pyudev/device/__init__.py
@@ -29,9 +29,3 @@ from ._device import Attributes
 from ._device import Device
 from ._device import Devices
 from ._device import Tags
-from ._errors import DeviceNotFoundAtPathError
-from ._errors import DeviceNotFoundByFileError
-from ._errors import DeviceNotFoundByNameError
-from ._errors import DeviceNotFoundByNumberError
-from ._errors import DeviceNotFoundError
-from ._errors import DeviceNotFoundInEnvironmentError
